rename folder name for local option store
